refactor(compliance): rename turbolens_security service to compliance_scanner
Step 1 of relocating the compliance scanner out of TurboLens. The
service was renamed when the CVE scanner was removed but the file
was left as 'turbolens_security' (a name that no longer means
anything: it's not a TurboLens feature anymore, and it doesn't do
security scanning). Renames:

- services/turbolens_security.py → services/compliance_scanner.py
- tests/services/test_turbolens_security.py → test_compliance_scanner.py
- tests/services/test_turbolens_security_compliance_dedup.py
  → test_compliance_scanner_dedup.py

Updates every import + module reference across api/v1/turbolens.py
and the rest of the test suite. Historical Alembic migrations keep
their original docstrings referencing the old name as period-correct
context.
